Single Function
This is a set of images that change background image on click. This is currently from a single function with each div id in an array and each image url in a separate array. a loop iterates through each array and then passes it through a function that changes the image background. only the array items need to be changed or added, not the function. However, I cannot get the images to return the original background on a second click, will work on this.